@@ -10,7 +10,7 @@ template<class T>
 static int num_digits(T number) {
   int digits = 0;
 
-  //if (number < 0) digits = 1; // remove this line if '-' counts as a digit
+  if (number < 0) digits = 1; // remove this line if '-' counts as a digit
 
   while (number) {
     number /= 10;
@@ -39,7 +39,9 @@ void lsms::print_dft_energy(const DFTEnergy &energy) {
   size = std::max(size, num_digits(static_cast<int> (energy.zero_point)));
   size = std::max(size, num_digits(static_cast<int> (energy.lsf)));
   size = std::max(size, num_digits(static_cast<int> (energy.madelung)));
-  size = std::max(size, num_digits(static_cast<int> (energy.shift)));
+  size = std::max(size, num_digits(static_cast<int> (energy.it_madelung)));
+  size = std::max(size, num_digits(static_cast<int> (energy.it_xc)));
+  size = std::max(size, num_digits(static_cast<int> (energy.mtz)));
   size = std::max(size, num_digits(static_cast<int> (energy.total)));
 
   size += 12;
@@ -59,9 +61,17 @@ void lsms::print_dft_energy(const DFTEnergy &energy) {
   std::printf("%-12s = %*.10f Ry\n", "XC", size, energy.xc);
   std::printf("%-12s = %*.10f Ry\n", "ZPE", size, energy.zero_point);
   std::printf("%-12s = %*.10f Ry\n", "LSF", size, energy.lsf);
+  std::printf("%-12s = %*.10f Ry\n", "Interstial", size, energy.u0);
 
-  std::printf("%-12s = %*.10f Ry\n", "Madelung", size, energy.madelung);
-  std::printf("%-12s = %*.10f Ry\n", "MT Zero", size, energy.shift);
+
+  std::printf("-------------------\n");
+
+  std::printf("%-12s = %*.10f Ry\n", "MT Madelung", size, energy.madelung);
+  std::printf("%-12s = %*.10f Ry\n", "IT Madelung", size, energy.it_madelung);
+  std::printf("%-12s = %*.10f Ry\n", "IT XC", size, energy.it_xc);
+  std::printf("%-12s = %*.10f Ry\n", "MTZ", size, energy.mtz);
+
+  std::printf("-------------------\n");
 
   std::printf("%-12s = %*.10f Ry\n", "Total energy", size, energy.total);
 
@@ -81,7 +91,7 @@ void lsms::globalSum(LSMSCommunication &comm, DFTEnergy &dft_energy) {
 
   MPI_Datatype dft_type;
 
-  constexpr int size = 14;
+  constexpr int size = 17;
 
   std::vector<MPI_Datatype> types(size, MPI_DOUBLE);
   std::vector<int> blocks(size, 1);
@@ -101,8 +111,12 @@ void lsms::globalSum(LSMSCommunication &comm, DFTEnergy &dft_energy) {
   MPI_Get_address(&dft_energy.xc, &displacements[9]);
   MPI_Get_address(&dft_energy.lsf, &displacements[10]);
   MPI_Get_address(&dft_energy.total, &displacements[11]);
+
   MPI_Get_address(&dft_energy.madelung, &displacements[12]);
-  MPI_Get_address(&dft_energy.shift, &displacements[13]);
+  MPI_Get_address(&dft_energy.it_madelung, &displacements[13]);
+  MPI_Get_address(&dft_energy.it_xc, &displacements[14]);
+  MPI_Get_address(&dft_energy.mtz, &displacements[15]);
+  MPI_Get_address(&dft_energy.u0, &displacements[16]);
 
   for (auto &value: displacements) {
     value -= base;
