@@ -120,7 +120,7 @@ public:
   Real clight;
   int maxlmax;
   LSMSGlobals global;
-  AngularMomentumIndices angularMomentumIndices;
+  Matrix<Real> dl_factor;
   EnergyContourParameters energyContour;
 
 // no. of Gaussian points for volume integration
@@ -135,6 +135,8 @@ public:
   Real volumeNorm;             // Volume renormalization factor
   Real volumeInterstitial;     // Total interstitial volume
   Real u0;                     // Contribution of the Muffin-tin zero potential to the Coulomb energy
+  Real u0MT;                   // Contribution of the Muffin-tin zero potential to the Coulomb energy just from MT
+  Real vmt;                    // Muffin-tin zero
   Real totalEnergy;            // Total energy
   //Real pressure;               // Pressure
 
