@@ -1,4 +1,6 @@
 /* -*- c-file-style: "bsd"; c-basic-offset: 2; indent-tabs-mode: nil -*- */
+#include "initializeAtom.hpp"
+
 #include <vector>
 #include <algorithm>
 #include <cstdio>
@@ -6,7 +8,6 @@
 #include "Real.hpp"
 #include "Matrix.hpp"
 #include "SingleSite/AtomData.hpp"
-#include "initializeAtom.hpp"
 #include "Core/calculateCoreStates.hpp"
 #include "Misc/integrateOneDim.hpp"
 
