@@ -77,33 +77,10 @@ T simpson_nonuniform(std::vector<T> x, std::vector<T> f, U length) {
 
 /**
  * Radial integrator for any mesh
- *
- * This method doesn't use any interpolations for the last point and assumes an
- * alight mesh. The mesh should be align to according to the
- * `generateRadialMesh` method
  */
 template <typename T>
 T radialIntegral(const std::vector<T> &integrand,
-                 const std::vector<T> &radial_mesh, const T r_sphere) {
-  // Index of element that is less then `r_sphere`
-  auto iter = std::upper_bound(radial_mesh.rbegin(), radial_mesh.rend(),
-                               r_sphere, std::greater<T>());
-
-  // Index of element that has radius greater then equal `r_sphere`
-  auto i_after_rs = std::distance(radial_mesh.begin(), iter.base());
-
-  auto end_point = i_after_rs + 1;
-
-
-  return simpson_nonuniform(radial_mesh, integrand, end_point);
-}
-
-/**
- * Radial integrator for any mesh
- */
-template <typename T>
-T radialIntegral(const std::vector<T> &integrand,
-                 const std::vector<T> &radial_mesh, int end) {
+                 const std::vector<T> &radial_mesh, std::size_t end) {
   auto result = simpson_nonuniform(radial_mesh, integrand, end);
 
   return result;
@@ -119,10 +96,10 @@ T radialIntegral(const std::vector<T> &integrand,
  */
 template <typename T>
 T radialIntegralDerivMesh(const std::vector<T> &val,
-                          const std::vector<T> &radial_mesh_deriv, int end) {
+                          const std::vector<T> &radial_mesh_deriv, std::size_t end) {
   T intval = 0.0;
-  int idx;
 
+  auto idx = end;
   for (idx = 0; idx < end - 2; idx += 2) {
     intval += 1.0 / 3.0 *
               (val[idx] * radial_mesh_deriv[idx] +
@@ -135,53 +112,6 @@ T radialIntegralDerivMesh(const std::vector<T> &val,
                      val[idx + 1] * radial_mesh_deriv[idx + 1]);
   };
   return intval;
-}
-
-template <typename T>
-T radialIntegral(const std::vector<T> &integrand,
-                 const std::vector<T> &radial_mesh, std::vector<T> &integral,
-                 const T r_sphere) {
-  // Index of element that is less then `r_sphere`
-  auto iter = std::upper_bound(radial_mesh.rbegin(), radial_mesh.rend(),
-                               r_sphere, std::greater<T>());
-
-  // Index of element that has radius greater then equal `r_sphere`
-  auto i_after_rs = std::distance(radial_mesh.begin(), iter.base());
-
-  // Create a shifted mesh so that the zeros point can be added
-  std::vector<T> extended_integrand(i_after_rs + 2);
-  std::vector<T> extended_r_mesh(i_after_rs + 2);
-
-  for (int i = 0; i < i_after_rs; ++i) {
-    extended_integrand[i + 1] = integrand[i];
-    extended_r_mesh[i + 1] = radial_mesh[i];
-  }
-
-  extended_integrand[0] = 0.0;
-  extended_r_mesh[0] = 0.0;
-
-  // Interpolate to the last value
-  auto lagrange =
-      integrand[i_after_rs - 1] * (r_sphere - radial_mesh[i_after_rs]) /
-          (radial_mesh[i_after_rs - 1] - radial_mesh[i_after_rs]) +
-      integrand[i_after_rs] * (r_sphere - radial_mesh[i_after_rs - 1]) /
-          (radial_mesh[i_after_rs] - radial_mesh[i_after_rs - 1]);
-
-  extended_integrand[i_after_rs + 1] = lagrange;
-  extended_r_mesh[i_after_rs + 1] = r_sphere;
-
-  auto extended_length = i_after_rs + 2;
-
-  auto result =
-      simpson_nonuniform(extended_r_mesh, extended_integrand, extended_length);
-
-  integral[0] = 0.0;
-  for (int i = 1; i < extended_length; ++i) {
-    integral[i] =
-        simpson_nonuniform(extended_r_mesh, extended_integrand, i + 1);
-  }
-
-  return result;
 }
 
 }  // namespace lsms
