@@ -7,16 +7,22 @@
 #include <complex>
 #include <vector>
 
+#include "Coeficients.hpp"
 #include "lattice_utils.hpp"
 #include "madelung_term.hpp"
 #include "monopole_madelung.hpp"
 
 lsms::MultipoleMadelung::MultipoleMadelung(LSMSSystemParameters &lsms,
                                            CrystalParameters &crystal,
-                                           LocalTypeInfo &local)
-    : num_atoms{crystal.num_atoms}, lmax{0}, local_num_atoms{local.num_local} {
+                                           LocalTypeInfo &local, int lmax)
+    : num_atoms{crystal.num_atoms},
+      lmax{lmax},
+      local_num_atoms{local.num_local} {
   auto r_brav = crystal.bravais;
   auto k_brav = crystal.bravais;
+
+  int kmax = (lmax + 1) * (lmax + 1);
+  int jmax = (lmax + 1) * (lmax + 2) / 2;
 
   // Scaling factors for the optimal truncation sphere
   scaling_factor = lsms::scaling_factor(crystal.bravais, lmax);
@@ -66,24 +72,40 @@ lsms::MultipoleMadelung::MultipoleMadelung(LSMSSystemParameters &lsms,
    * Calculate Madelung matrix for monopoles (lmax = 0)
    */
   auto omega = lsms::omega(r_brav);
-
+  auto alat =
+      scaling_factor * std::cbrt(3.0 * omega / (4.0 * M_PI * num_atoms));
 
   // Zero terms
   auto term0 = -M_PI * eta * eta / omega;
 
+  /**
+   * Resize and set values to zero madelung Matrix and also multipole madelung
+   * matrix
+   */
+
   for (auto j = 0; j < local.num_local; j++) {
     local.atom[j].madelungMatrix.resize(crystal.num_atoms);
+    std::fill_n(local.atom[j].madelungMatrix.begin(), crystal.num_atoms, 0.0);
+  }
+
+  if (jmax > 1) {
+    for (auto j = 0; j < local.num_local; j++) {
+      local.atom[j].multipoleMadelung.resize(kmax, crystal.num_atoms);
+      local.atom[j].multipoleMadelung = std::complex<double>(0.0, 0.0);
+    }
   }
 
   double timeLoopSpace = MPI_Wtime();
 
-  // Introduced for smaller objects
+  // Smaller object for positons
   auto position = crystal.position;
 
-  #pragma omp parallel for collapse(2) firstprivate(nknlat, nrslat, scaling_factor, position, knlatsq, knlat, rslat, rslatsq) default(shared)
+#pragma omp parallel for collapse(2)                                       \
+    firstprivate(nknlat, nrslat, scaling_factor, position, knlatsq, knlat, \
+                 rslat, rslatsq, term0) shared(local, eta, omega, jmax, kmax, lmax, alat) \
+                 default(none)
   for (int atom_i = 0; atom_i < num_atoms; atom_i++) {
     for (int local_i = 0; local_i < local_num_atoms; local_i++) {
-
       std::vector<double> aij(3);
       double r0tm;
       int ibegin;
@@ -94,7 +116,7 @@ lsms::MultipoleMadelung::MultipoleMadelung(LSMSSystemParameters &lsms,
       // a_ij in unit of a0
       for (int idx = 0; idx < 3; idx++) {
         aij[idx] = position(idx, atom_i) / scaling_factor -
-            position(idx, global_i) / scaling_factor;
+                   position(idx, global_i) / scaling_factor;
       }
 
       // Real space terms: first terms
@@ -116,6 +138,24 @@ lsms::MultipoleMadelung::MultipoleMadelung(LSMSSystemParameters &lsms,
       // Madelung matrix contribution
       local.atom[local_i].madelungMatrix[atom_i] =
           (term1 + term2 + r0tm + term0) / scaling_factor;
+
+      if (jmax > 1) {
+        // {l,m} = 0,0
+        local.atom[local_i].multipoleMadelung(0, atom_i) =
+            local.atom[local_i].madelungMatrix[atom_i] * Y0inv;
+
+        std::vector<Complex> dlm(kmax, std::complex<double>(0, 0));
+
+        lsms::dlsum(aij, rslat, nrslat, ibegin, knlat, nknlat, omega, lmax, eta,
+                    dlm);
+
+        // Higher order terms
+        for (int kl = 1; kl < kmax; kl++) {
+          auto l = AngularMomentumIndices::lofk[kl];
+          local.atom[local_i].multipoleMadelung(kl, atom_i) =
+              dlm[kl] * std::pow(alat / scaling_factor, l) / scaling_factor;
+        }
+      }
     }
   }
 
@@ -127,16 +167,71 @@ lsms::MultipoleMadelung::MultipoleMadelung(LSMSSystemParameters &lsms,
     std::printf("Time: %16s %lf\n", "Loop:", timeLoopSpace);
   }
 
+  /*
+   * Prefactors for the transformation of reduced Madelung constants
+   *
+   *  C^{l',m'}_{l,m},(l+l')(m-m')
+   *
+   */
+
+  if (jmax > 1) {
+    // Variable
+    lsms.dl_factor.resize(kmax, jmax);
+
+    std::vector<double> factmat(lmax + 1);
+    factmat[0] = 1.0;
+    for (int l = 1; l <= lmax; l++) {
+      factmat[l] = factmat[l - 1] / (2.0 * l + 1.0);
+    }
+
+    /**
+     * Reference: Zabloudil S. 218
+     */
+
+#pragma omp parallel for collapse(2) \
+    shared(lsms) firstprivate(factmat, jmax, kmax) default(none)
+    for (int jl_pot = 0; jl_pot < jmax; jl_pot++) {
+      for (int kl_rho = 0; kl_rho < kmax; kl_rho++) {
+
+        // Static variables are shared by default
+        auto l_pot = AngularMomentumIndices::lofj[jl_pot];
+        auto kl_pot = AngularMomentumIndices::kofj[jl_pot];
+        auto l_rho = AngularMomentumIndices::lofk[kl_rho];
+
+        auto l_sum = l_pot + l_rho;
+
+        // m2 - m1 = m3  otherwise zero this is always true
+        // l1 + l2 + l3 needs to be even
+        int j3 = l_sum / 2;
+
+        lsms.dl_factor(kl_rho, jl_pot) =
+            GauntCoeficients::cgnt(j3, kl_pot, kl_rho) * factmat[l_pot] *
+            factmat[l_rho];
+
+      }
+    }
+  }
 }
 
-double lsms::MultipoleMadelung::getScalingFactor() const {
+__attribute__((unused)) double lsms::MultipoleMadelung::getScalingFactor()
+const {
   return scaling_factor;
 }
 
-double lsms::MultipoleMadelung::getRsCut() const { return rscut; }
+__attribute__((unused)) double lsms::MultipoleMadelung::getRsCut() const {
+  return rscut;
+}
 
-double lsms::MultipoleMadelung::getKnCut() const { return kncut; }
+__attribute__((unused)) double lsms::MultipoleMadelung::getKnCut() const {
+  return kncut;
+}
 
-std::vector<int> lsms::MultipoleMadelung::getKnSize() const { return k_nm; }
+__attribute__((unused)) std::vector<int> lsms::MultipoleMadelung::getKnSize()
+const {
+  return k_nm;
+}
 
-std::vector<int> lsms::MultipoleMadelung::getRsSize() const { return r_nm; }
+__attribute__((unused)) std::vector<int> lsms::MultipoleMadelung::getRsSize()
+const {
+  return r_nm;
+}
