@@ -4,14 +4,15 @@
 
 #include "calculateMultipoleMadelung.hpp"
 
+#include "Misc/Coeficients.hpp"
 #include "MultipoleMadelung.hpp"
 
 void calculateMultiMadelungMatrices(LSMSSystemParameters &lsms,
                                     CrystalParameters &crystal,
-                                    LocalTypeInfo &local) {
+                                    LocalTypeInfo &local, int lmax) {
   if (lsms.global.iprint >= 0) {
     std::printf("Madelung calculations!\n");
   }
 
-  lsms::MultipoleMadelung madelung(lsms, crystal, local);
+  lsms::MultipoleMadelung madelung(lsms, crystal, local, lmax);
 }
