@@ -5,5 +5,5 @@
 
 //getvmt.f in LSMS 1
 void getvmt(LSMSSystemParameters lsms, AtomData &atom, CrystalParameters &crystal, Real qsub[], int &mytype, Real &vmt,
-            Real &vmt1, Real &u0);
+            Real &vmt1, Real &u0, Real &u0MT);
 
